docs: update architecture and developer guide
