+++ collectors/openai_feed_test.go
@@ -1,4 +1,4 @@
-package exporter
+package collectors
 
 import (
 	"net/http"
