+++ collectors/testdata/aws_multi_item.rss
