+++ collectors/testdata/aws_outage.rss
