+++ collectors/testdata/azure_issue.rss
