+++ collectors/testdata/gcp_resolved_then_update.atom
