+++ collectors/testdata/gcp_update_resolved.atom
