+++ collectors/testdata/openai_resolved.atom
