+++ collectors/workers.go
@@ -1,12 +1,12 @@
-package exporter
+package collectors
 
 import (
 	"context"
 	"sync"
 )
 
-// startWorkers launches a monitor goroutine for each service and returns a
-// WaitGroup that waits for all workers to exit.
+// StartWorkers launches a monitor goroutine for each service and
+// returns a WaitGroup that waits for all workers to exit.
 func StartWorkers(ctx context.Context, services []ServiceFeed) *sync.WaitGroup {
 	var wg sync.WaitGroup
 	for _, svc := range services {
